fix/ fiz alterações na composição do componente username, dividindo em dois componentes devido a quantidade de utilização na aplicação.
