+++ src/App.tsx
@@ -1,5 +1,4 @@
 
-
 const App = () => {
   return <></>;
 };
