+++ src/components/InicialsName/style.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const ContainerInitial = styled.div`
+
+    border-radius: 50%;  
+    left: 277px;
+    top: 23.5px;
+    display: flex;
+    align-items: center;
+    justify-content: center;    
+
+
+h2{
+    font-family: Inter; 
+    font-weight: 700;
+    letter-spacing: 0em;
+    text-align: left;
+    color: var(--color-white-fixed);
+}
+`
