+++ src/components/Name/style.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const ContainerName = styled.div`
+
+h2{
+    font-family: Inter;   
+    font-weight: 400;
+    line-height: 28px;
+    letter-spacing: 0em;
+    text-align: left;
+    color: var(--color-grey-2);
+    margin-left: 5px;
+}
+
+`
